Refactor some TreeView handlers to supporting files to clean things up
